# pkg/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "REST API for ShvarkPay",
        "title": "Shvark API Gateway",
        "contact": {},
        "version": "1.0"
    },
    "host": "http://localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/payments/in/h2h": {
            "post": {
                "description": "Create new Pay-In using host-to-host method",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payments"
                ],
                "summary": "Create new H2H Pay-In",
                "parameters": [
                    {
                        "description": "pay-in info",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.CreateH2HPayInRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/response.CreateH2HPayInResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.BadRequestErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.NoBankDetailsErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/payments/in/h2h/{id}": {
            "get": {
                "description": "Get host-to-host pay-in order info",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payments"
                ],
                "summary": "Get H2h Pay-in info",
                "parameters": [
                    {
                        "type": "string",
                        "description": "order id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.GetH2HPayInInfoResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.BadRequestErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/payments/in/h2h/{id}/arbitrage/info": {
            "get": {
                "description": "Get dispute info by disputeID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payments"
                ],
                "summary": "Get info about dispute",
                "parameters": [
                    {
                        "type": "string",
                        "description": "dispute id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.GetPayInArbitrageInfoResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/payments/in/h2h/{id}/arbitrage/link": {
            "post": {
                "description": "Opent dispute by order id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payments"
                ],
                "summary": "Open dispute for given order",
                "parameters": [
                    {
                        "type": "string",
                        "description": "order ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "dispute description",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_LavaJover_shvark-api-gateway_internal_delivery_http_dto_payment_request.CreateDisputeRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_LavaJover_shvark-api-gateway_internal_delivery_http_dto_payment_response.CreateDisputeResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/payments/in/h2h/{id}/cancel": {
            "post": {
                "description": "Cancel Pay in order",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payments"
                ],
                "summary": "Cancel Pay In order",
                "parameters": [
                    {
                        "type": "string",
                        "description": "order ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.CancelPayInResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_LavaJover_shvark-api-gateway_internal_delivery_http_dto_payment_request.CreateDisputeRequest": {
            "type": "object",
            "properties": {
                "amount_fiat": {
                    "type": "number"
                },
                "proof_url": {
                    "type": "string"
                },
                "reason": {
                    "type": "string"
                }
            }
        },
        "github_com_LavaJover_shvark-api-gateway_internal_delivery_http_dto_payment_response.CreateDisputeResponse": {
            "type": "object",
            "properties": {
                "dispute_id": {
                    "type": "string"
                },
                "expires_at": {
                    "type": "string"
                }
            }
        },
        "github_com_LavaJover_shvark-api-gateway_internal_delivery_http_dto_payment_response.Dispute": {
            "type": "object",
            "properties": {
                "dispute_id": {
                    "type": "string"
                },
                "dispute_reason": {
                    "type": "string"
                },
                "dispute_status": {
                    "type": "string"
                },
                "order_id": {
                    "type": "string"
                },
                "proof_url": {
                    "type": "string"
                }
            }
        },
        "handlers.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "invalid data"
                }
            }
        },
...
        "request.CreateH2HPayInRequest": {
            "type": "object",
            "properties": {
                "amount_fiat": {
                    "type": "number"
                },
                "callback_url": {
                    "type": "string"
                },
                "client_id": {
                    "type": "string"
                },
                "currency": {
                    "type": "string"
                },
                "ftd": {
                    "type": "boolean"
                },
                "merchant_id": {
                    "type": "string"
                },
                "merchant_order_id": {
                    "type": "string"
                },
                "payment_system": {
                    "type": "string"
                },
                "shuffle": {
                    "type": "integer"
                },
                "ttl": {
                    "type": "string"
                }
            }
        },
...
        "response.BadRequestErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "response.CancelPayInResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "response.CreateH2HPayInResponse": {
            "type": "object",
            "properties": {
                "amount_crypto": {
                    "type": "number"
                },
                "amount_fiat": {
                    "type": "number"
                },
                "callback_url": {
                    "type": "string"
                },
                "crypto_rub_rate": {
                    "type": "number"
                },
                "currency": {
                    "type": "string"
                },
                "expires_at": {
                    "type": "integer"
                },
                "merchant_order_id": {
                    "type": "string"
                },
                "order_id": {
                    "type": "string"
                },
                "payment_details": {
                    "$ref": "#/definitions/response.PaymentDetails"
                },
                "payment_system": {
                    "type": "string"
                },
                "recalculated": {
                    "type": "boolean"
                },
                "status": {
                    "type": "string"
                },
                "tpay_link": {
                    "type": "string"
                }
            }
        },
...
        "response.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "response.GetH2HPayInInfoResponse": {
            "type": "object",
            "properties": {
                "amount_crypto": {
                    "type": "number"
                },
                "amount_fiat": {
                    "type": "number"
                },
                "callback_url": {
                    "type": "string"
                },
                "crypto_rub_rate": {
                    "type": "number"
                },
                "currency": {
                    "type": "string"
                },
                "expires_at": {
                    "type": "integer"
                },
                "merchant_order_id": {
                    "type": "string"
                },
                "order_id": {
                    "type": "string"
                },
                "payment_details": {
                    "$ref": "#/definitions/response.PaymentDetails"
                },
                "payment_system": {
                    "type": "string"
                },
                "recalculated": {
                    "type": "boolean"
                },
                "status": {
                    "type": "string"
                },
                "tpay_link": {
                    "type": "string"
                }
            }
        },
...
        "response.GetPayInArbitrageInfoResponse": {
            "type": "object",
            "properties": {
                "dispute": {
                    "$ref": "#/definitions/github_com_LavaJover_shvark-api-gateway_internal_delivery_http_dto_payment_response.Dispute"
                }
            }
        },
...
        "response.NoBankDetailsErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "response.PaymentDetails": {
            "type": "object",
            "properties": {
                "bank": {
                    "description": "User-friendly",
                    "type": "string"
                },
                "bank_id": {
                    "description": "In SBP system",
                    "type": "string"
                },
                "bank_name": {
                    "type": "string"
                },
                "card_number": {
                    "type": "string"
                },
                "owner": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}